requirements.txt
flask==2.2.3
werkzeug==2.2.3
moviepy==1.0.3
numpy==1.24.3
pillow==9.5.0
google-generativeai==0.3.1
python-dotenv==1.0.0
gtts==2.3.2
flask-wtf==1.1.1
Werkzeug==2.2.3
pydub==0.25.1 